requirements.txt
google-genai
streamlit
fastapi
uvicorn
requests
